Merged PR 113: wimg: clean memory before writting our firmware, using the max size for early bootloader
This PR fix #284 by full erasing the memory of the LORIXOne before writing our firmware.

As LORIX Programming tool is dedicated for Wifx usage, it only write the needed portion of the firmware, with some artefact remaining. In our case, we need to have a full erase, which is done by "flash-mix" the old flashing tool sam-ba.

To do so, in the limit of the feature offered by LORIX Programming tool, we write the at91 bootloader "until the end of the memory", thus 0 values are written after it on the whole memory (as done by sam-ba), and then we wrote back our partitions and firmware.

See related issue for more information and flash logs.

Related work items: #284