Merged PR 117: Add support for Wifx L1 hardware
This PR add the support of Wifx L1 hardware on OpenLorix.
Arolys will be treated in a dedicated PR

Wifx L1 support new RF chip SX1302 (SX1301 for Lorix One) that involve to include a new driver.
The hardware has an usb-c port that require to upgrade linux version previously used for Lorix One hardware

One of the constraints was to continue the support of Lorix One. The update of OpenWrt was to be avoided to not have to readapt and retest the patches made for Lorix One

## Linux version
To support usb-c, linux version was upgraded from 4.9 to 4.14. The choice of the version is hardware dependent to avoid involving other hardware(see makefile)
Linux version 4.14 is the highest version possible of the actual OpenWrt version.
Wifx use linux version 5.4 so patches.
The existing patches had to be adapted to go from version 4.9 to 4.14. The patches provided by Wifx had to be adapted to go from version 4.5 to 4.14

## Bootstrap version
Wifx use at91bootstrap-3.9.1, we use at91bootstrap-3.8.10. Wifx patche to support L1 has been adapted to our version

## U-boot version
Wifx use version v2018.09. We use linux4sam_5.8. No need to upgrade u-boot version. Wifx patche to support L1 has been adapted to our version

## Elvexys-oss-seed
Packages had to be adapted to support the L1 and its SX1302 chip but will be the subject of another PR.

Related work items: #332